+++ ecommerce_project/myapp/serializers/product_serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 from ecommerce_project.myapp.models import Category, Company, SellerUser, Product
 from ecommerce_project.myapp.serializers.seller_serializers import SellerOutputSerializer
-from ecommerce_project.myapp.serializers.OtherSerializers import CategorySerializer
+from ecommerce_project.myapp.serializers.other_serializers import CategorySerializer
 from ecommerce_project.myapp.serializers.company_serializers import CompanyOutputSerializer
