+++ ecommerce_project/myapp/views/__init__.py
@@ -1,6 +1,6 @@
 from .user_views import *
 from .other_views import *
-from .product_related_views import *
+from .cart_views import *
 from .product_views import *
 from .review_views import *
 from .order_views import *
