+++ ecommerce_project/myapp/views/cart_views.py
@@ -1,12 +1,9 @@
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from ecommerce_project.myapp.models import Review, Order, OrderLine, Cart, CartLine
-from ecommerce_project.myapp.serializers import ReviewOutputSerializer, ReviewInputSerializer, CartOutputSerializer, \
+from ecommerce_project.myapp.models import Cart, CartLine
+from ecommerce_project.myapp.serializers import  CartOutputSerializer, \
     CartInputSerializer, CartLineOutputSerializer, CartLineInputSerializer
-from ecommerce_project.myapp.serializers.order_serializers import  \
-    OrderOutputSerializer, OrderInputSerializer, \
-    OrderLineOutputSerializer, OrderLineInputSerializer
 
 class CartListNCreateAPIView(APIView):
     # permission_classes = [IsAuthenticated]
